merging quasar code into webwork3 (#23)
* initial checkin of vue3-quasar code.

* more checkin of code.

* switching to more quasar-standard setup

* update install instructions.

* Vue3 quasar (#16)

* linting cleanup of markdown documents

* deleting to rename file.

* Create WeBWorK3.pm

* add perl tests to github action

* Change to superlinter v4 and more cleanup

* make perl-test-script executable

* error in linter github action

* bump superlinter to 4.2.2

* switching to more quasar-standard setup

* update install instructions.

* adding course_settings to db and tests

* fixes to course settings

* added test to user sets

* a few bug fixes

* some changes to get devel::cover working

* testing github action

* fix for github action.

* fix for github action.

* add build_db script

* another go

* another go

* yet another go

* another go

* added git to docker

* remove token

* reinstated secret

* minor change to trigger test

* change secret token name

* added badges

* minor edit to fire action

* turn off upload for git action

* added template views and rudimentary calendar

* work on the Settings view

* add routes to course settings

* fixing some linting errors

* fixed minor render error

* preliminary checkin of Course Settings

* initial checking of module code for user, problem_sets

* fixed an error fetching problem sets

* more updates to problem set manager

* Changed the way that dates are read in/processed in tests

* minor fixes after merge

* Lots of updates to functionality (#20)

* linting cleanup of markdown documents

* deleting to rename file.

* Create WeBWorK3.pm

* add perl tests to github action

* Change to superlinter v4 and more cleanup

* make perl-test-script executable

* error in linter github action

* bump superlinter to 4.2.2

* switching to more quasar-standard setup

* update install instructions.

* adding course_settings to db and tests

* fixes to course settings

* added test to user sets

* a few bug fixes

* some changes to get devel::cover working

* testing github action

* fix for github action.

* fix for github action.

* add build_db script

* another go

* another go

* yet another go

* another go

* added git to docker

* remove token

* reinstated secret

* minor change to trigger test

* change secret token name

* added badges

* turn off upload for git action

* added template views and rudimentary calendar

* work on the Settings view

* add routes to course settings

* fixing some linting errors

* fixed minor render error

* preliminary checkin of Course Settings

* initial checking of module code for user, problem_sets

* fixed an error fetching problem sets

* more updates to problem set manager

* minor fixes after merge

* more work on ClasslistManager

* updates to user db functionality

* updates to classlist manager

* add ability to use mariadb/mysql database for tests

* initi checkin of some admin views

* more updates on adding a course

* More updates on adding a course

* update getGlobalUser

* Updates to AddCourse

* Add vuex-persistedstate to save the vuex session state to session (#21)

storage.  This makes page refresh work (and has other possibile uses).

Remove the webwork3 url prefix from the code, and instead put that into
the publicPath option in quasar.conf.js.

Add some eslint rule for syntax.  These are rules that I prefer, but are
up for discussion.  One thing that this does is disables console.log's
for the production build.  Note that these should also not be committed
to git (they can be committed commented out if the must be there).  All
existing console.log's have been removed from the code.

Add the .editorconfig file.  This is honored by many text editors by
default.  Some require a plugin (for example vim).  See
https://editorconfig.org/ for details.

Add stylelint to the build system.  It only emits errors for the
production build, and not for the development build.

Clean up the build system.  Mainly remove most of the comments added by
the quasar cli.  All of those are well documented on the quasar website.
Also remove things from the build system that we will not be using like
ssr and pwa.  These can be added back if they are used.

Remove the .quasar directory from git, and add it to the ignore file.
That directory is generated by the build, and its contents are different
for the development build and the production build.  It should not be in
the repository.

Switch from using the user's email address to using the user's login id
(a username).  I am not sure why the email address was used.  We need to
discuss this.

Add a method for deployment to the README.md file.

There are a few other minor improvements to the ui, but most of the
other changes are clean up and fixes for the lintian rules.

* more changes after merge

* Fixed a number of linting and other errors after merge

* Fixed linting and other post-merge errors

* minor fixes for linting

* updates to docker file and other changes

* importing more compatible version of vuejs.

Co-authored-by: Glenn Rice <47527406+drgrice1@users.noreply.github.com>