Fix the database to work with an actual database management system like (#33)
mysql or mariadb.  There were two problems that made the database code
fail with those systems.  First, columns of type text can not be part of
a primary key or index.  So those columns were switched to varchar type.
Second, the JSON columns were not correctly coded.  That could have been
fixed, but the DBIx::Class::InflateColumn::Serializer::JSON package
handles this nicely, so that was used.  On Ubuntu you can install the
package libdbix-class-inflatecolumn-serializer-perl.

Also restrucure the database settings in the webwork3.dist.yml file a
bit.  There is no need to distinguish the cases for sqlite, mariadb, or
mysql, and so a `database` config variable is not needed.  Just set the
`database_dsn`.  The `database_user` and `database_password` variables
are ignored if sqlite is used.

Also, the build_db.pl file was made executable for convenience.

Note that the docker image (drgrice1/webwork3) used by the unit test
workflow has been updated for these changes.  The updated
docker/webwork3.dockerfile file that was used for this build is included
in this pull request.