Api clean up, add session cookie settings, and use axios correctly (#39)
* Remove all of the Mojolicious routes that are not part of the api.
Those are not needed with the Quasar/Vue approach.  Get rid of the
template files used for that.  Also rename Mojolicious.pm to Hooks.pm.
We don't need a file that is named for the library we are using. Name it
for what it is.

Set up the session for the api with cookie settings from the
configuration.

Use subroutine signatures since we have enabled the Mojolicious
signatures feature.  This is only done in lib/WeBWorK3.pm and
lib/WeBWorK3/Controller/Login.pm at this point, but should be done
consistently in all perl files if we are going to use the feature.
Perlcritic is not up to date on this and thinks these are prototypes.
So a .perlcriticrc file is added to disable that warning.

Furthermore, the perlcritic severity is set to level 4.  The resulting
warnings that stem from this are fixed.

Use axios the correct way according to the Quasar documentation.  This
is also better for our usage.  The api base url is part of the
boot/axios setup.

Remove the "@" alias added in the tsconfig.json file.  That overrides
quasar's aliases and is inconsistent with the rest of the webpack build.

Restructure .eslintrc.js so that typescript rules are only applied to
typescript files, and not javascript files.

* Update the login unit tests to test via the api only.

* fix typo

* Fix the perlcritic warnings in the test files.  I forgot to check these
files before.  Thanks for pointint this out @drdrew42.

Co-authored-by: K. Andrew Parker <kparker@citytech.cuny.edu>