Fix the typescript unhandled diagnostic that occurs when the quasar dev (#46)
server is run while the dist folder from the build is present.  This is
caused by an incorrect exclusion in the @quasar/app/tsconfig-preset.json
file.  In that file the directory is excluded as "/dist".  The forward
slash does not work here.  So the exclusion is overridden in our
tsconfig.json file.