Add a minimal part of Bootstrap 5 that does not conflict with Quasar, and use that to get several things working in problems.
Scaffolding works.

MathQuill works with the tooltips again.

A modified version of the image view javascript and style is added, and
that works.

Basic styles for inputs are added so that those look better (assuming
mathquill is not enabled for the input).

Much of this requires some changes to the renderer and pg.  These
changes are in the ww3-rendering-work branch of my fork of the renderer.
Make sure to run `git submodules update` after switching to that branch
to also get my branch of pg with the necessary changes there.