Add logging to the Problem.vue file for errors that occur in loading a problem.
Also update the code to work for multiple problems rendered on the same
page.

Move the MathJax typesetting code into the mathjax-config.ts file and
improve that to update the global MathJax startup promise.  This is from
the MathJax documentation, and is required for multiple asynchronous
calls to work correctly.  Without this if multiple problems are on the
same page there are issues.

When processing and executing script tags in the loaded problem html,
don't process math/tex script tags for MathJax.

If the `sourceFilePath` (previously `file`) is set when the Problem.vue
component is mounted, then immediately load the problem.  The `watch`
method does not see this.