Update the internationalization setup.
Add configuration options to the app i18n object to make it work with
the vue composition api.  Also add a fallback locale.

Lazy load the locales from yaml files.  We really don't want to load all
languages by default.

Make the language selector show the language names instead of the
iso language code.  Note that the displayed language name is always in
its own language regardless of the selected language.  This is how it
should be.

Finally, add a few more translations.