Fix some issues with problem rendering.
First, add the problem buttons back so that problems can be tested again.

Remove the pg javascript.  That will now be loaded by PG itself.  Note
that to test this you will need to use the
ww3-rendering-work-plus-more-new branch of my renderer fork.  Make sure
to update your git submodules also as the lib/PG branch used is
different.  PG requests the basic javascript and css needed.  It is
assumed that third party javascript and css that is needed is provided
(jquery, bootstrap, and mathjax and technically jquery-ui but we don't
do that yet or maybe won't).

Make sure that a problems renderer_params are set from the
_request_fields.  This is needed to ensure a proper clone via the
`clone` method.

Fix a bug in the parseParams method.  That should not modify the passed
in _params directly.  This is called by the constructor, and so by the
clone method which calls the constructor for the clone, and that ends up
potentically modifying the original problem that was being cloned.