Fix permissions and require authentication for api routes
Remove the ignore_permissions option and use permissions in unit test.
The permissions are fixed so that they actually work.

Also require authentication to access all api routes except the login
and logout routes.

Fix the cookie settings in the conf file.  The samesite setting does not
need the quotes, and the secure setting needs to be a numeric 0 or 1.
The string true did work because that is truthy in perl.  However, it is
misleading because the string false is also truthy.