Clean up a few things.
There is no need to import things in `boot` with the `src` prefix.
There is an alias for `boot` provided by Quasar.

Please start paying attention to the order of imports.  It is important
for understanding the code and parallels between related code.  Generally
think top down with imports in terms of the application.  Things
imported from third party libraries generally come first, and then think
in terms of application generality.  For example Vue is the most broad
library, Quasar depends on that, etc.  So Vue imports first, then Quasar
imports, and so on.

Also start adding more space for better separation of code.
Particularly between methods.

Generally, work on consistency in your coding style.

These are things that are extremely important in team projects.  This
helps other developers understand what you are doing.  It also helps you
understand what you are doing.