Add two new perltidy options.
First add the extended continuation indentation option. I added this to
the webwork2 and pg .perltidyrc files, and it does a better job of
indenting for nested structures.  This rule responsible for the changes
to the other files in this pull request.

Second add the option to disable vertical alignment of qw quotes.  Since
this option is added, the workflow version of perltidy can now be
allowed to increase.  I left the lock though, and just moved it to the
newest version.