run test-stores now spawns  morbo.  (#91)
This changes the way that npm run test-stores runs. Instead of requiring morbo to be started, this will spawn it before running the jest script.

Additionally, morbo is started in 'test' mode which loads conf/ww3-dev.yaml instead of conf/webwork3.yaml.

Other thoughts about this. I set the port this runs as 3333. Perhaps have an environment variable to set this.

Also, I haven't put this in the GitHub action. It might run now that morbo is running.