Clean up the conf file handling.
Rename ww3-dev.dist.yml to webwork3-test.dist.yml.  The ww3 abreviation
is not needed and is inconsistent in naming with the other conf files.

You can now create a separate webwork3-dev.yml file for use in
development.  There is no associated dist file for this.  Just make
whatever changes you want to either of the webwork3*.dist.yml files.

Delete the we_b_wor_k3.yml link.  Just load the appropriate conf file in
conf directly.

Completely eliminate the WW3_ROOT environment variable.  There is no
need for that.  Instead make use of Mojo::Home.  It is made for this.
If you are in an area of the code that has access to the app, then just
use the home helper (in a controller this can be accessed via
$c->app->home).  Elsewhere in the code use Mojo::Home as is done in
lib/WeBWorK3/Utils/Settings.pm.

Also fix the usage of the exception classes in
lib/WeBWorK3/Utils/Settings.pm.  My editor complains about the way those
were imported there.  Furthermore it was entirely inconsistent with the
way the exception classes are imported elsewhere.  There is a mess of
ways that imports are done in the perl files.  This needs further clean
up.  In some places qw// is used, in others qw() is used (both of those
occur in lib/WeBWorK3/Utils/Settings.pm), in yet others a list of
individually quoted imports is used (see lib/DB/WithParams.pm).

@pstaapb: Find some consistency in your coding style.