Massive overhaul of the units tests.
The unit tests no longer use the conf/webwork3-test.dist.yml or
conf/webwork3-test.yml configuration files.  Instead the configuration
is hardcoded into the unit test runners defined in t/lib/DBSubtest.pm
and bin/dev_scripts/test_pinia_stores.pl.  The point is that unit tests
should not depend on some modifiable configuration.  The
conf/webwork3-test.dist.yml file has been deleted.  You can copy
conf/webwork3.dist.yml to conf/webwork3-dev.yml for development use.
For morbo or the `bin/dev_scripts/build_sample_db.pl` script (previously
`t/db/build_db.pl`) the file conf/webwork3-dev.yml will be used if it
exists.

The sample data for the database is now in JSON files instead of CSV
files.  This allows for structured data with valid types.  This means
that all of the crap that was done to convert the data from the CSV
files into structured data with valid types is not needed.  In perl
decode_json does what is needed, and in javascript the JSON files can
just be imported and typescript correctly infers the types.  Note that
eventually it would be nice to do what is done in t/db/001_courses.t for
more of the tests.  That test no longer uses the sample data in the files
at all.  Most of the tests really don't use much from the sample data.
Furthermore, the tests can made much more precise and transparent in
this way.

The unit tests each deploy a temporary database instance, and load the
sample data automatically.  In addition, the tests can be run with all
three supported databases (sqlite, postgres, and mysql).  By default the
tests are only run with an in memory sqlite instance.  For all of the
tests if the environment variable WW3_TEST_ALL_DBS is set then the tests
will instead each be run three times.  Once with each of the supported
databases.  For postgres the Test::PostgreSQL package is used to create
a temporary postgres instance.  For mysql a local modified version of
the Test::mysqld package (t/lib/TestMysqld.pm) is used.  The upstream
Test::mysqld package has issues in that it must be run as root.  Note
that on the other hand the Test::PostgreSQL package can not be run as
root.  Note that for these tests to work you must have postgres and
mysql (or mariadb) installed. However, they do not need to be running on
the system.  Since the database instances are temporary and each test
uses its own instance, there is no need to be concerned with putting the
database back to the way it was.  It will be deleted when the test is
completed anyway.

The bin/dev_scripts/test_pinia_stores bash script has been replaced with
the bin/dev_scripts/test_pinia_stores.pl perl script.  That script runs
a Mojolicious server daemon and creates a webwork3 api app instance with
a provided test configuration (no longer in the lib/WeBWorK3.pm file),
and then runs the jest pinia store tests.  The server port can be
configured either with a command line option or via the WW3_TEST_PORT
environment variable.  The test runs by default only with the in memory
sqlite database, but if the -ta (or --test-all-dbs) option is given or
the WW3_TEST_ALL_DBS environment variable is set, then the test will run
three times, once with each of the three databases.

Note that the Github action runs all unit tests with all databases.
Previously the pinia stores tests were not run by the action, but now
they are (and with all three databases).

Converting the CSV files to JSON files and working with that data has
illuminated some big problems that need to be fixed.

First, the data returned by the webwork3 api methods currently returns
data with extremely poor data structure. The data returned by the api
methods should not be flattened in the way that it is currently.  This
is the biggest issue and most pressing.  Both the api and the client
side stores will need to be fixed.

Second, the "get all" methods need to be removed.  These are abolutely
unnecessary, shouldn't exist, and should not be tested.  For example,
there is no reason to get all problems in the database.  Problems should
only be obtained in the context that they have meaning.  That is in the
context of the set they belong to.  It may be that these were added for
the purpose of the unit tests to check that things were put back the way
that the were in the database.  If so, these still shouldn't have been
in the code.  These methods were all essentially just the DBIx::Class
search method which could have been used directly.  In addition those
checks are now not needed with this change.  The getAllProblemSets
method was another of these. That is removed.

Note that the unit tests now run all all pushes.  This is useful before
making a pull request to see that the unit tests pass.  The code
coverage is only uploaded when a pull request is merged into main in any
case, so there is no reason not to have these run.